backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn==0.27.1
starlette>=0.27.0
pydantic>=2.0,<3.0
python-multipart==0.0.9
python-dotenv==1.0.1

# HTTP and Requests
httpx>=0.26.0,<0.29.0  # Required by supabase 2.15.1
requests==2.31.0

# Google Cloud
google-cloud-storage==2.14.0
google-generativeai==0.3.2

# Document Processing
PyMuPDF==1.23.26
pytesseract==0.3.10
Pillow==10.2.0
tqdm==4.66.2
tiktoken==0.6.0  # Required for token counting

# Database
supabase==2.15.1  # Latest stable version
postgrest>0.19,<1.1  # Required by supabase 2.15.1

# LangChain Ecosystem
langchain==0.1.9
langchain-community>=0.0.21,<0.1.0  # Required by langchain
langchain-core>=0.1.0  # Required by langchain
langchain-text-splitters>=0.0.1  # Required by langchain
langchain-google-genai==0.0.5
langchain-google-vertexai==0.0.1
